refactor(analytics): extract shared clickhouse record decoders (adr-066)
Extract the five byte-identical JSONEachRow decoders (asNumber, asNullableNumber, asStringArray, asStringMap, asNullableString) shared by the trace_analytics and evaluation_analytics read-back repositories into src/server/clickhouse/recordDecode.ts. The eval repo keeps its column-specific asNullableMs / asNullableBool local until a second consumer appears.

Also document the pre-00055 latched-boolean read-back limitation: HasAnnotation and a user-overridden TraceName cannot fully round-trip because their backing columns (AnnotationIds / TraceNameUserOverridden) did not exist before 00055, so a pre-00055 row can downgrade the flag on a post-miss rewrite. Corrects the mixed-deploy self-heal note in the migration header and the fold doc.
